+++ cogs/help.py
@@ -1,6 +1,6 @@
 import discord
 from discord.ext import commands
-from discord.commands import slash_command, Option
+from discord.commands import slash_command
 
 class Help(commands.Cog):
 
@@ -14,12 +14,13 @@ class Help(commands.Cog):
 		help_embed.set_author(name='BRUHbot', icon_url='https://images-ext-2.discordapp.net/external/0OavdISaaL7vL2nD4sL6VgEz_ow6i42Pdo_jWdHfcYQ/%3Fsize%3D1024/https/cdn.discordapp.com/avatars/745294860839420034/4f4fa33ae47e49a5f7aca65d92bcaf75.webp')
 		help_embed.add_field(name='Category: Moderation', value='`ban, kick, purge, selfrole, giverole`',  inline=False)
 		help_embed.add_field(name='Category: Security', value='`obfs, deobfs, piglatin, genpass, dcpl`')
-		help_embed.add_field(name='Category: Normal Stuff', value='`sup, givefood, pfp, latency, luckcheck, nocontext, ask, gdlevelrate, aboutwhale, mts, feedback`')
-		help_embed.add_field(name='Category: Media', value='`smile, mrkrabsbruh, shrimp, dababy, latityintro, thelord, chickennugget, phishe, cat, hampter, kittyreview, cursedimg, fumo`')
+		help_embed.add_field(name='Category: Normal Stuff', value='`sup, givefood, pfp, latency, luckcheck, nocontext, ask, gdlevelrate, aboutwhale, mts, feedback, define`')
+		help_embed.add_field(name='Category: Media', value='`smile, mrkrabsbruh, shrimp, dababy, latityintro, thelord, chickennugget, phishe, cat, hampter, kittyreview, cursedimg, fumo, watermelon, lagh, gamer, skeleton, doorknob, willi, rapbattlehistory, tvmp4, truck, sauce`')
 		help_embed.add_field(name='Category: amogus (pls dont try this)', value='`sus, kill`')
-		help_embed.add_field(name='Category: Economy', value='`daily, balance, quest, share, shop`')
+		help_embed.add_field(name='Category: Economy', value='`daily, balance, quest, share, shop -> (list|buy)`')
 		help_embed.add_field(name='Category: Game', value='`rps, guessnum`')
-		help_embed.add_field(name='Category: Music', value='`queue, play, stop, pause, resume, ytsearch`')
+		help_embed.add_field(name='Category: Music', value='`play, ytsearch, join, leave`')
+		help_embed.add_field(name='Category: Anime', value='`anime -> (describe|search)`')
 		help_embed.set_footer(text='Im sleepy...')
 		await ctx.respond(embed=help_embed)